+|xrpc|client绑定channel
