+++ README.md
@@ -30,5 +30,5 @@ open source RPC framework
 3. 超时熔断
 
 ### 🎯 下一阶段开发工作
-1. client对象和channel对象进行绑定，而不是BootStrap 
+1. client对象和channel对象进行绑定，进行存活检测,长时间没有进行调用的channel进行关闭
 2. 注册中心整合nacos
