+++ xrpc-spring-boot-starter/src/main/java/com/xzq/spring/XrpcAutoConfiguration.java
@@ -1,5 +1,6 @@
 package com.xzq.spring;
 
+import com.xzq.client.XrpcMessageHandler;
 import com.xzq.client.proxy.ProxyFactory;
 import com.xzq.register.RedisRegister;
 import com.xzq.register.Register;
@@ -9,9 +10,16 @@ import com.xzq.server.factory.ProviderFactory;
 import com.xzq.spring.properties.RegisterProperties;
 import com.xzq.spring.properties.ServerProperties;
 import com.xzq.spring.properties.XrpcProperties;
+import com.xzq.xrpc.remoting.codec.MessageCodec;
+import com.xzq.xrpc.remoting.codec.ProtocolFrameDecoder;
 import com.xzq.xrpc.remoting.protocol.XrpcProtocol;
 import com.xzq.xrpc.remoting.protocol.XrpcProtocolV1;
+import io.netty.bootstrap.Bootstrap;
+import io.netty.channel.ChannelInitializer;
+import io.netty.channel.ChannelOption;
 import io.netty.channel.nio.NioEventLoopGroup;
+import io.netty.channel.socket.SocketChannel;
+import io.netty.channel.socket.nio.NioSocketChannel;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -44,6 +52,28 @@ public class XrpcAutoConfiguration {
     @Bean("workerGroup")
     public NioEventLoopGroup workerGroup() {
         return new NioEventLoopGroup();
+    }
+
+    @Bean
+    public Bootstrap bootstrap() {
+        Bootstrap bootstrap = new Bootstrap();
+
+        bootstrap
+                .group(workerGroup())
+                .channel(NioSocketChannel.class)
+                .option(ChannelOption.AUTO_READ, true)
+                .handler(new ChannelInitializer<SocketChannel>() {
+                    @Override
+                    protected void initChannel(SocketChannel socketChannel) throws Exception {
+                        //半包粘包解码器
+                        socketChannel.pipeline().addLast(new ProtocolFrameDecoder());
+                        //消息编解码器
+                        socketChannel.pipeline().addLast(new MessageCodec(xrpcProtocol()));
+                        //消息处理器
+                        socketChannel.pipeline().addLast(new XrpcMessageHandler());
+                    }
+                });
+        return bootstrap;
     }
 
     @Bean
@@ -81,7 +111,7 @@ public class XrpcAutoConfiguration {
 
     @Bean
     public ProxyFactory proxyFactory() {
-        return new ProxyFactory(xrpcProtocol());
+        return new ProxyFactory(xrpcProtocol(), bootstrap());
     }
 
     @Bean
